+++ src/SMAPI/Framework/Reflection/ReflectedMethod.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 namespace StardewModdingAPI.Framework.Reflection
 {
     /// <summary>A method obtained through reflection.</summary>
-    internal class ReflectedMethod : IPrivateMethod, IReflectedMethod
+    internal class ReflectedMethod : IReflectedMethod
     {
         /*********
         ** Properties